requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
